--- x/storage/keeper/grpc_query_params_test.go
@@ -1,22 +1,0 @@
-package keeper_test
-
-import (
-	"testing"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/stretchr/testify/require"
-
-	testkeeper "github.com/bnb-chain/greenfield/testutil/keeper"
-	"github.com/bnb-chain/greenfield/x/storage/types"
-)
-
-func TestParamsQuery(t *testing.T) {
-	keeper, _, ctx := testkeeper.StorageKeeper(t)
-	wctx := sdk.WrapSDKContext(ctx)
-	params := types.DefaultParams()
-	keeper.SetParams(ctx, params)
-
-	response, err := keeper.Params(wctx, &types.QueryParamsRequest{})
-	require.NoError(t, err)
-	require.Equal(t, &types.QueryParamsResponse{Params: params}, response)
-}
