+++ x/storage/keeper/grpc_query_test.go
@@ -1,17 +1,41 @@
 package keeper_test
 
 import (
+	storetypes "cosmossdk.io/store/types"
+	"github.com/cosmos/cosmos-sdk/testutil"
+	moduletestutil "github.com/cosmos/cosmos-sdk/types/module/testutil"
+	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
+	"github.com/cosmos/cosmos-sdk/x/mint"
 	"testing"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/stretchr/testify/require"
 
-	testkeeper "github.com/bnb-chain/greenfield/testutil/keeper"
+	"github.com/bnb-chain/greenfield/x/storage/keeper"
 	"github.com/bnb-chain/greenfield/x/storage/types"
 )
 
+func makeKeeper(t *testing.T) (*keeper.Keeper, sdk.Context) {
+	encCfg := moduletestutil.MakeTestEncodingConfig(mint.AppModuleBasic{})
+	key := storetypes.NewKVStoreKey(types.StoreKey)
+	testCtx := testutil.DefaultContextWithDB(t, key, storetypes.NewTransientStoreKey("transient_test"))
+
+	k := keeper.NewKeeper(
+		encCfg.Codec,
+		key,
+		&types.MockAccountKeeper{},
+		&types.MockSpKeeper{},
+		&types.MockPaymentKeeper{},
+		&types.MockPermissionKeeper{},
+		&types.MockCrossChainKeeper{},
+		authtypes.NewModuleAddress(types.ModuleName).String(),
+	)
+
+	return k, testCtx.Ctx
+}
+
 func TestParamsQuery(t *testing.T) {
-	keeper, _, ctx := testkeeper.StorageKeeper(t)
+	keeper, ctx := makeKeeper(t)
 	wctx := sdk.WrapSDKContext(ctx)
 	params := types.DefaultParams()
 	keeper.SetParams(ctx, params)
