+++ x/storage/keeper/params_test.go
@@ -5,12 +5,11 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	testkeeper "github.com/bnb-chain/greenfield/testutil/keeper"
 	"github.com/bnb-chain/greenfield/x/storage/types"
 )
 
 func TestGetParams(t *testing.T) {
-	k, _, ctx := testkeeper.StorageKeeper(t)
+	k, ctx := makeKeeper(t)
 	params := types.DefaultParams()
 
 	k.SetParams(ctx, params)
